Dismantle _format/utils.py into multiple files
